Controler 0.12
So as of now i only need a way to seperate the textfiles into diffrent lines and to read one line after another. When that is done i think it would be good enough that i only need to figure out how to make it send the signals and in what format it should send the signals to the CNC machine.